requirements.txt
pypdf
langchain
chromadb 
pytest
boto3
langchain-chroma
streamlit
tavily-python
